refactor : api client와 server 구분
