+++ apis/groups/index.ts
@@ -1,4 +1,5 @@
-export * from './apis';
+export * from './apis.client';
+export * from './apis.server';
 export * from './keys';
 export * from './mutations';
 export * from './queries';
